requirements.txt
Flask~=2.0.2
PyJWT~=2.3.0
flask-sqlalchemy~=2.5.1
flask_cors~=3.0.10
Werkzeug~=2.0.2